requirements.txt
python-fasthtml>=0.6.0
pandas>=1.5.0
matplotlib>=3.6.0
pytest>=7.0.0
